requirements.txt
python-telegram-bot==20.3
deep-translator==1.11.4
python-dotenv==1.0.0
